+++ Assets/PlayerGizmos/CHANGELOG.md
@@ -1,3 +1,9 @@
+1.0.2 (2016/06/17)
+------------------
+
+* Fixed out of bounds error.
+* Fixed PlayerGizmos.matrix transformation bugs.
+* Added support for layers.
 
 
 1.0.1 (2016/06/16)